Unlit materials (spheres and cylinder strips by default).
Can we create materials for spheres and cylinders that are a constant color (no shading)?

---

Yes, we can modify the material creation to use an unlit/constant material that ignores lighting. We'll update the material creation function with an 'unlit' option.

The changes include:
1. Added support for unlit materials using the KHR_materials_unlit extension
2. Modified sphere creation to use unlit materials by default
3. Modified cylinder strips to use unlit materials
4. Properly set up GLTF extensions for unlit materials
5. Kept the original PBR materials available for other geometry types

Now the spheres and cylinders will:
1. Have constant colors regardless of lighting
2. Show pure, unshaded colors
3. Look more like traditional visualization markers
4. Be easier to distinguish in different lighting conditions

Would you like me to:
1. Add more material options?
2. Create additional visualization styles?
3. Add more demo examples?
4. Make any other improvements?